+++ CHANGELOG.md
@@ -1,7 +1,8 @@
-## [unreleased]
+## [6.11.5] (2023-02-27)
 
 ### Fix
 - replace grpc-reflection-js with grpc-js-reflection-client
+- fix order of log output in cli (#644)
 
 ## [6.11.4] (2023-02-25)